countylevel tooltip update
